MediaViewer: Fix cut video player controls
Previously video player controls were getting cut off when displayed gif was taller than screen. This was caused by incorrect calculation of margins and height for video player.

This commit fixes the issue by changing the calculations to work with video player's parent size and padding which results in dimension that are correct and also work for split-screen.