Move clamping of max swipe distance to RecyclerSwipeListener
If a function has restrictions on parameter values, I think the onus of correcting invalid values should be on the caller. The function can throw an error on invalid values, but shouldn't correct them.

Additionally, function parameters are final in Kotlin. The last commit was updating the parameter value of SwipeableLayout#setSwipeTranslation() which wasn't future foolproof.